back-rewrite/api/requirements.txt
django
yt-dlp
pygame
daphne
ytmusicapi
django-cleanup
Twisted[tls,http2]
django-ninja
orjson
mutagen
discord.py[voice]
requests
json5
